requirements.txt
django==6.0.3
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
inflection==0.5.1
psycopg==3.2.12
uritemplate==4.2.0
